[BE] REFACTOR: API 분리
