+++ backend/src/relation/relation.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { FriendsController } from './friends.controller';
+import { BlockController } from './block.controller';
+
+@Module({
+  controllers: [FriendsController, BlockController],
+})
+export class RelationModule {}
