+++ backend/src/stat/stat.module.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { StatController } from './stat.controller';
+
+@Module({
+  controllers: [StatController],
+})
+export class StatModule {}
